// packages/coin-tron/src/protobuf/tron.d.ts
const $protobuf = _m0

/** Namespace protocol. */
export namespace protocol {
// ...
    namespace Transaction {

        /** Properties of a Contract. */
        interface IContract {

            /** Contract type */
            type?: (protocol.Transaction.Contract.ContractType|null);

            /** Contract parameter */
            parameter?: (google.protobuf.IAny|null);

            /** Contract provider */
            provider?: (Uint8Array|null);

            /** Contract ContractName */
            ContractName?: (Uint8Array|null);

            /** Contract PermissionId */
            PermissionId?: (number|null);
        }
// ...
        namespace Contract {

            /** ContractType enum. */
            enum ContractType {
                AccountCreateContract = 0,
                TransferContract = 1,
                TransferAssetContract = 2,
                VoteAssetContract = 3,
                VoteWitnessContract = 4,
                WitnessCreateContract = 5,
                AssetIssueContract = 6,
                WitnessUpdateContract = 8,
                ParticipateAssetIssueContract = 9,
                AccountUpdateContract = 10,
                FreezeBalanceContract = 11,
                UnfreezeBalanceContract = 12,
                WithdrawBalanceContract = 13,
                UnfreezeAssetContract = 14,
                UpdateAssetContract = 15,
                ProposalCreateContract = 16,
                ProposalApproveContract = 17,
                ProposalDeleteContract = 18,
                SetAccountIdContract = 19,
                CustomContract = 20,
                CreateSmartContract = 30,
                TriggerSmartContract = 31,
                GetContract = 32,
                UpdateSettingContract = 33,
                ExchangeCreateContract = 41,
                ExchangeInjectContract = 42,
                ExchangeWithdrawContract = 43,
                ExchangeTransactionContract = 44,
                UpdateEnergyLimitContract = 45,
                AccountPermissionUpdateContract = 46,
                ClearABIContract = 48
            }
        }
// ...
    }
// ...
}

/** Namespace google. */
export namespace google {

    /** Namespace protobuf. */
    namespace protobuf {

        /** Properties of an Any. */
        interface IAny {

            /** Any type_url */
            type_url?: (string|null);

            /** Any value */
            value?: (Uint8Array|null);
        }

        /** Represents an Any. */
        class Any implements IAny {

            /**
             * Constructs a new Any.
             * @param [properties] Properties to set
             */
            constructor(properties?: google.protobuf.IAny);

            /** Any type_url. */
            public type_url: string;

            /** Any value. */
            public value: Uint8Array;

            /**
             * Creates a new Any instance using the specified properties.
             * @param [properties] Properties to set
             * @returns Any instance
             */
            public static create(properties?: google.protobuf.IAny): google.protobuf.Any;

            /**
             * Encodes the specified Any message. Does not implicitly {@link google.protobuf.Any.verify|verify} messages.
             * @param message Any message or plain object to encode
             * @param [writer] Writer to encode to
             * @returns Writer
             */
            public static encode(message: google.protobuf.IAny, writer?: $protobuf.Writer): $protobuf.Writer;

            /**
             * Encodes the specified Any message, length delimited. Does not implicitly {@link google.protobuf.Any.verify|verify} messages.
             * @param message Any message or plain object to encode
             * @param [writer] Writer to encode to
             * @returns Writer
             */
            public static encodeDelimited(message: google.protobuf.IAny, writer?: $protobuf.Writer): $protobuf.Writer;

            /**
             * Decodes an Any message from the specified reader or buffer.
             * @param reader Reader or buffer to decode from
             * @param [length] Message length if known beforehand
             * @returns Any
             * @throws {Error} If the payload is not a reader or valid buffer
             * @throws {$protobuf.util.ProtocolError} If required fields are missing
             */
            public static decode(reader: ($protobuf.Reader|Uint8Array), length?: number): google.protobuf.Any;

            /**
             * Decodes an Any message from the specified reader or buffer, length delimited.
             * @param reader Reader or buffer to decode from
             * @returns Any
             * @throws {Error} If the payload is not a reader or valid buffer
             * @throws {$protobuf.util.ProtocolError} If required fields are missing
             */
            public static decodeDelimited(reader: ($protobuf.Reader|Uint8Array)): google.protobuf.Any;

            /**
             * Verifies an Any message.
             * @param message Plain object to verify
             * @returns `null` if valid, otherwise the reason why it is not
             */
            public static verify(message: { [k: string]: any }): (string|null);

            /**
             * Creates an Any message from a plain object. Also converts values to their respective internal types.
             * @param object Plain object
             * @returns Any
             */
            public static fromObject(object: { [k: string]: any }): google.protobuf.Any;

            /**
             * Creates a plain object from an Any message. Also converts values to other types if specified.
             * @param message Any
             * @param [options] Conversion options
             * @returns Plain object
             */
            public static toObject(message: google.protobuf.Any, options?: $protobuf.IConversionOptions): { [k: string]: any };

            /**
             * Converts this Any to JSON.
             * @returns JSON object
             */
            public toJSON(): { [k: string]: any };
        }
    }
}